Move skiplist_handler.py to a common place
`skiplist_handler.py` is used by the analyze and server part of the
CodeChecker too. To prepare source repository split up we move it to
a common place.
